+++ json_serializable/test/default_value/default_value.checked.dart
@@ -22,7 +22,10 @@ const _intValue = 42;
 dvi.DefaultValue fromJson(Map<String, dynamic> json) =>
     _$DefaultValueFromJson(json);
 
-@JsonSerializable()
+@JsonSerializable(
+  anyMap: true,
+  checked: true,
+)
 class DefaultValue implements dvi.DefaultValue {
   @JsonKey(defaultValue: true)
   bool fieldBool;